add logging service for campaign
