+++ Projects/App/Sources/AppDelegate+DI.swift
@@ -48,7 +48,7 @@ extension ApolloClient {
         } else {
             let hApollo = self.createClient()
             let paymentService = hPaymentClientOctopus()
-            let hCampaignsService = hCampaingsServiceOctopus()
+            let hCampaignsService = hCampaingsClientOctopus()
             let networkClient = NetworkClient()
             let messagesClient = FetchMessagesClientOctopus()
             let sendMessage = SendMessagesClientOctopus()
@@ -78,7 +78,7 @@ extension ApolloClient {
                 Dependencies.shared.add(module: Module { () -> FileUploaderClient in networkClient })
                 Dependencies.shared.add(module: Module { () -> AdyenService in networkClient })
                 Dependencies.shared.add(module: Module { () -> hPaymentClient in paymentService })
-                Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
+                Dependencies.shared.add(module: Module { () -> hCampaignClient in hCampaignsService })
                 Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
                 Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
                 Dependencies.shared.add(module: Module { () -> FetchContractsService in fetchContractsService })
@@ -105,7 +105,7 @@ extension ApolloClient {
                 Dependencies.shared.add(module: Module { () -> FileUploaderClient in networkClient })
                 Dependencies.shared.add(module: Module { () -> AdyenService in networkClient })
                 Dependencies.shared.add(module: Module { () -> hPaymentClient in paymentService })
-                Dependencies.shared.add(module: Module { () -> hCampaignsService in hCampaignsService })
+                Dependencies.shared.add(module: Module { () -> hCampaignClient in hCampaignsService })
                 Dependencies.shared.add(module: Module { () -> hFetchClaimService in hFetchClaimService })
                 Dependencies.shared.add(module: Module { () -> hClaimFileUploadService in networkClient })
                 Dependencies.shared.add(module: Module { () -> FetchContractsService in fetchContractsService })
