+++ Projects/Payment/Sources/Service/CampaignService/CampaignClient.swift
@@ -1,0 +1,24 @@
+import Foundation
+import hCore
+import hGraphQL
+
+public protocol hCampaignClient {
+    func remove(codeId: String) async throws
+    func add(code: String) async throws
+}
+
+enum CampaignError: Error {
+    case userError(message: String)
+    case notImplemented
+}
+
+extension CampaignError: LocalizedError {
+    var errorDescription: String? {
+        switch self {
+        case let .userError(message):
+            return message
+        case .notImplemented:
+            return L10n.General.errorBody
+        }
+    }
+}
